LayoutTests/http/tests/local/blob/send-hybrid-blob-expected.txt

CONSOLE MESSAGE: line 134: ReferenceError: Can't find variable: WebKitBlobBuilder
FAIL: Timed out waiting for notifyDone to be called

Test for building blobs with multiple files combined by BlobBuilder and sending them via XMLHttpRequest. (This test requires eventSender.beginDragWithFiles)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* BlobBuilder.append(file) - single file
PASS successfullyParsed is true

TEST COMPLETE
